Replaced ng-bootstrap with ngx-bootstrap

+++ src/renderer/app/components/players/players.component.ts
@@ -1,11 +1,11 @@
 import { Component, OnInit, ElementRef, ViewChild } from '@angular/core';
-import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
+import { BsModalService } from 'ngx-bootstrap/modal';
 
 import { PlayerService } from '../../services/player.service';
 import { Player } from '../../models/player.model';
 import { RoleService } from '../../services/role.service';
-import IPlayerRole from '../../models/roles/playerrole.i';
-import { SettingsService } from '../../services/settings.service';
+import { PlayerModalComponent, PlayerModalResult } from '../modals/player/player-modal.component';
+import { ConfirmModalComponent } from "../modals/confirm/confirm-modal.component";
 
 @Component({
   selector: 'cc-players',
@@ -14,25 +14,10 @@ import { SettingsService } from '../../services/settings.service';
 })
 export class PlayersComponent implements OnInit {
 
-  @ViewChild("confirmDeleteModal")
-  private _confirmDeleteModal: ElementRef;
-
-  @ViewChild("addPlayerModal")
-  private _addPlayerModal: ElementRef;
-
-  @ViewChild("confirmDeadModal")
-  private _confirmDeadModal: ElementRef;
-
-  public createPlayerFirstName: string = "";
-  public createPlayerLastName: string = "";
-  public createPlayerAmount: number = this._settingsService.playerStartingAmount;
-  public createPlayerRole: IPlayerRole;
-
   constructor(
     private readonly _playerService: PlayerService,
     private readonly _roleService: RoleService,
-    private readonly _modalService: NgbModal,
-    private readonly _settingsService: SettingsService
+    private readonly _modalService: BsModalService
   ) { }
 
   ngOnInit() {
@@ -42,48 +27,46 @@ export class PlayersComponent implements OnInit {
     return this._playerService.players;
   }
 
-  public get roles(): IPlayerRole[] {
-    return this._roleService.availableRoles;
-  }
-
   public createPlayer(): void {
-    this._modalService.open(this._addPlayerModal).result.then(() => {
-      const newPlayer = this._playerService.createPlayer(this.createPlayerFirstName, this.createPlayerLastName, this.createPlayerAmount);
+    this._modalService.show(PlayerModalComponent).content.result.subscribe(
+      (player: PlayerModalResult) => {
+        const newPlayer = this._playerService.createPlayer(player.firstName, player.lastName, player.startingAmount);
 
-      if (!!this.createPlayerRole) {
-        this._roleService.assignRole(this.createPlayerRole.name, newPlayer);
-      }
-
-      this.clearCreatePlayerInputs();
-    }, () => {
-      this.clearCreatePlayerInputs();
-    }); 
-  }
-
-  private clearCreatePlayerInputs(): void {
-    this.createPlayerFirstName = "";
-    this.createPlayerLastName = "";
-    this.createPlayerRole = null;
+        if (!!player.role) {
+          this._roleService.assignRole(player.role.name, newPlayer);
+        }
+      });
   }
 
   public deletePlayer(player: Player) {
-    this._modalService.open(this._confirmDeleteModal).result.then(() => {
-      this._playerService.deletePlayer(player);
-    }, () => {});
+    this._modalService.show(ConfirmModalComponent, {
+      initialState: {
+        title: "Delete Player",
+        text: `Are you sure you wish to delete ${player.name}?`
+      }
+    }).content.result.subscribe(
+      (result: boolean) => {
+        if (!result) return;
+        this._playerService.deletePlayer(player);
+        this._roleService.unassignPlayer(player);
+    });
   }
 
   public setDead(player: Player) {
-    this._modalService.open(this._confirmDeadModal).result.then(() => {
-      player.setDead();
-    }, () => {});
+    this._modalService.show(ConfirmModalComponent, {
+      initialState: {
+        title: "Player Dead",
+        text: `Are you sure you wish to set ${player.name} to dead?`
+      }
+    }).content.result.subscribe(
+      (result: boolean) => {
+        if (!result) return;
+        player.setDead();
+    });
   }
 
   public getPlayerRole(player: Player): string {
     const role = this._roleService.getRoleForPlayer(player);
     return !!role ? role.name : "";
-  }
-
-  public setPlayerRole(player: Player, role: string) {
-    this._roleService.assignRole(role, player);
   }
 }
